requirements.txt
demjson3
mypy
pylint
pyrefly
ruff
types-PyYAML
yamllint
yq
